+++ protected/humhub/modules/user/models/forms/EditGroupForm.php
@@ -16,17 +16,68 @@ use Yii;
  */
 class EditGroupForm extends Group
 {
+    public const TYPE_NORMAL = 'normal';
+    public const TYPE_SUBGROUP = 'subgroup';
+    public string $type;
+    public $subgroups;
+    public $parent;
+
     public $managerGuids = [];
     public $defaultSpaceGuid = [];
     public $updateSpaceMemberships = false;
 
+    /**
+     * @inheritdoc
+     */
+    public function afterFind()
+    {
+        parent::afterFind();
+
+        if (!isset($this->type)) {
+            $this->type = $this->parent_group_id === null ? self::TYPE_NORMAL : self::TYPE_SUBGROUP;
+        }
+
+        switch ($this->type) {
+            case self::TYPE_NORMAL:
+                $this->subgroups = Group::find()
+                    ->where(['parent_group_id' => $this->id])
+                    ->select('id')
+                    ->column();
+                break;
+            case self::TYPE_SUBGROUP:
+                $this->parent = [$this->parent_group_id];
+                break;
+        }
+    }
+
     public function rules()
     {
         $rules = parent::rules();
-        $rules[] = [['name'], 'required'];
+        $rules[] = [['name', 'type'], 'required'];
         $rules[] = [['updateSpaceMemberships'], 'boolean'];
         $rules[] = [['managerGuids', 'show_at_registration', 'show_at_directory', 'defaultSpaceGuid'], 'safe'];
+        $rules[] = [['subgroups', 'parent'], 'safe'];
+        $rules[] = [['type'], 'validateTypeGroups'];
         return $rules;
+    }
+
+    public function validateTypeGroups()
+    {
+        if ($this->type === self::TYPE_SUBGROUP && empty($this->parent[0])) {
+            $this->addError('parent', Yii::t('AdminModule.user', 'Parent group is required!'));
+        }
+
+        if ($this->isNewRecord) {
+            return;
+        }
+
+        if (is_array($this->subgroups) && in_array($this->id, $this->subgroups)) {
+            $this->addError('subgroups', 'Subgroup cannot be same current group!');
+        }
+
+        if (is_array($this->parent) && in_array($this->id, $this->parent)) {
+            $this->addError('parent', 'Parent group cannot be same current group!');
+        }
     }
 
     public function attributeLabels()
@@ -34,6 +85,9 @@ class EditGroupForm extends Group
         return array_merge(parent::attributeLabels(), [
             'defaultSpaceGuid' => Yii::t('AdminModule.space', 'Default Space(s)'),
             'updateSpaceMemberships' => Yii::t('AdminModule.space', 'Update Space memberships also for existing members.'),
+            'type' => Yii::t('AdminModule.user', 'Group Type'),
+            'subgroups' => Yii::t('AdminModule.user', 'Subgroup(s)'),
+            'parent' => Yii::t('AdminModule.user', 'Parent Group'),
         ]);
     }
 
@@ -71,6 +125,19 @@ class EditGroupForm extends Group
             $groupSpaces->save();
         }
 
+        Group::updateAll(['parent_group_id' => null], ['parent_group_id' => $this->id]);
+        switch ($this->type) {
+            case self::TYPE_NORMAL:
+                $this->updateAttributes(['parent_group_id' => null]);
+                if (is_array($this->subgroups) && $this->subgroups !== []) {
+                    Group::updateAll(['parent_group_id' => $this->id], ['id' => $this->subgroups]);
+                }
+                break;
+            case self::TYPE_SUBGROUP:
+                $this->updateAttributes(['parent_group_id' => $this->parent]);
+                break;
+        }
+
         parent::afterSave($insert, $changedAttributes);
     }
 
@@ -100,5 +167,13 @@ class EditGroupForm extends Group
                 }
             }
         }
+    }
+
+    public function getTypeOptions(): array
+    {
+        return [
+            self::TYPE_NORMAL => Yii::t('AdminModule.user', 'Normal'),
+            self::TYPE_SUBGROUP => Yii::t('AdminModule.user', 'Subgroup'),
+        ];
     }
 }